Move modules into package and add packaging
