+++ README.md
@@ -1,2 +1,12 @@
 # OCI Deep Research
-This repository contains the code for a concept/demo around Deep Research, using OCI
+This repository contains the code for a concept/demo around Deep Research using OCI.
+
+## Installation
+
+Install the package in editable mode to use the nodes and workflow:
+
+```bash
+pip install -e .
+```
+
+This command installs the `oci_deep_research` package so you can import its classes and functions.
