+++ oci_deep_research/classify_topic_node.py
@@ -7,11 +7,11 @@ Author: L. Saetta
 from langchain.prompts import PromptTemplate
 from langchain_core.runnables import Runnable
 from langchain_core.messages import HumanMessage
-from report_state import ReportState
-from prompts import PROMPT_TEMPLATE_TOPIC
-from model_factory import get_chat_model
-from utils import get_console_logger, extract_json_from_text
-from config import DEBUG
+from .report_state import ReportState
+from .prompts import PROMPT_TEMPLATE_TOPIC
+from .model_factory import get_chat_model
+from .utils import get_console_logger, extract_json_from_text
+from .config import DEBUG
 
 logger = get_console_logger(name="ClassifyTopicNode")
