+++ oci_deep_research/generate_section_node.py
@@ -7,10 +7,10 @@ Author: L. Saetta
 from langchain.prompts import PromptTemplate
 from langchain_core.runnables import Runnable
 from langchain_core.messages import HumanMessage
-from report_state import ReportState
-from prompts import PROMPT_TEMPLATE_GENERATE
-from model_factory import get_chat_model
-from utils import get_console_logger
+from .report_state import ReportState
+from .prompts import PROMPT_TEMPLATE_GENERATE
+from .model_factory import get_chat_model
+from .utils import get_console_logger
 
 logger = get_console_logger(name="GenerateSectionNode")
