+++ oci_deep_research/plan_report_node.py
@@ -10,11 +10,11 @@ Author: L. Saetta
 from langchain.prompts import PromptTemplate
 from langchain_core.runnables import Runnable
 from langchain_core.messages import HumanMessage
-from report_state import ReportState
-from prompts import PROMPT_TEMPLATE_PLAN
-from model_factory import get_chat_model
-from utils import get_console_logger, extract_json_from_text
-from config import DEBUG, DEFAULT_MAX_SECTIONS, DEFAULT_MIN_SECTIONS
+from .report_state import ReportState
+from .prompts import PROMPT_TEMPLATE_PLAN
+from .model_factory import get_chat_model
+from .utils import get_console_logger, extract_json_from_text
+from .config import DEBUG, DEFAULT_MAX_SECTIONS, DEFAULT_MIN_SECTIONS
 
 logger = get_console_logger(name="PlanReportNode")
