+++ oci_deep_research/retrieve_info_node.py
@@ -7,11 +7,11 @@ Author: L. Saetta
 from langchain.prompts import PromptTemplate
 from langchain_core.runnables import Runnable
 from langchain_core.messages import HumanMessage
-from report_state import ReportState
-from prompts import PROMPT_TEMPLATE_SEARCH
-from model_factory import get_chat_model
-from utils import get_console_logger
-from config import DEBUG, SEARCH_INFO_MODEL_ID
+from .report_state import ReportState
+from .prompts import PROMPT_TEMPLATE_SEARCH
+from .model_factory import get_chat_model
+from .utils import get_console_logger
+from .config import DEBUG, SEARCH_INFO_MODEL_ID
 
 logger = get_console_logger(name="RetrieveInfoNode")
