+++ oci_deep_research/review_report_node.py
@@ -7,11 +7,11 @@ Author: L. Saetta
 from langchain.prompts import PromptTemplate
 from langchain_core.runnables import Runnable
 from langchain_core.messages import HumanMessage
-from report_state import ReportState
-from prompts import PROMPT_TEMPLATE_REVIEW
-from model_factory import get_chat_model
-from utils import get_console_logger
-from config import FINAL_REPORT_MAX_TOKENS
+from .report_state import ReportState
+from .prompts import PROMPT_TEMPLATE_REVIEW
+from .model_factory import get_chat_model
+from .utils import get_console_logger
+from .config import FINAL_REPORT_MAX_TOKENS
 
 logger = get_console_logger(name="ReviewReportNode")
