+++ pyproject.toml
@@ -1,0 +1,15 @@
+[build-system]
+requires = ["setuptools>=61"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "oci-deep-research"
+version = "0.1.0"
+description = "OCI Deep Research agent demo"
+authors = [{name = "L. Saetta"}]
+readme = "README.md"
+requires-python = ">=3.8"
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["oci_deep_research*"]
